Added custom middleware
